// src/app/stock/interfaces/stock.interface.ts
export interface StockInformation {
    Symbol:                     string;
    AssetType:                  string;
    Name:                       string;
    Description:                string;
    CIK:                        string;
    Exchange:                   string;
    Currency:                   string;
    Country:                    string;
    Sector:                     string;
    Industry:                   string;
    Address?:                    string;
    FiscalYearEnd?:              string;
    LatestQuarter?:              Date;
    MarketCapitalization?:       string;
    EBITDA?:                     string;
    PERatio?:                    string;
    PEGRatio?:                   string;
    BookValue?:                  string;
    DividendPerShare?:           string;
    DividendYield?:              string;
    EPS?:                        string;
    RevenuePerShareTTM?:         string;
    ProfitMargin?:               string;
    OperatingMarginTTM?:         string;
    ReturnOnAssetsTTM?:          string;
    ReturnOnEquityTTM?:          string;
    RevenueTTM?:                 string;
    GrossProfitTTM?:             string;
    DilutedEPSTTM?:              string;
    QuarterlyEarningsGrowthYOY?: string;
    QuarterlyRevenueGrowthYOY?:  string;
    AnalystTargetPrice?:         string;
    TrailingPE?:                 string;
    ForwardPE?:                  string;
    PriceToSalesRatioTTM?:       string;
    PriceToBookRatio?:           string;
    EVToRevenue?:                string;
    EVToEBITDA?:                 string;
    Beta?:                       string;
    "52WeekHigh"?:               string;
    "52WeekLow"?:                string;
    "50DayMovingAverage"?:       string;
    "200DayMovingAverage"?:      string;
    SharesOutstanding?:          string;
    DividendDate?:               Date;
    ExDividendDate?:             Date;
}